requirements.txt
streamlit
openai
python-dotenv
beautifulsoup4
markdownify
streamlit_pills
trafilatura
